// src/types/create/types.d.ts
export interface DeliveryLocation {
 lgt: string,
 lat: string,
 note: string,
 district: string,
 town: string,
 street: string,
 country: string
}